feat: Color style and tone icons in customization header
This change updates the style and tone customization header to display the
font icons with their respective style and color.

The `CollapsibleSection` component has been updated to accept an `iconColors`
prop, which is an array of color strings. This allows passing the colors
of the style and tone icons.

The `MainPage` now passes the colors of the selected style and tone to the
`CollapsibleSection` component.

Additionally, the `useLocalStorage` hook has been refactored into its own
file for better code organization. Test files have been updated to
accommodate these changes and to fix existing issues.
